feat: move ReadGenesis to common
